+++ example/counter.ts
@@ -1,7 +1,6 @@
 import { KingWorld } from '../src'
 
 new KingWorld()
-	// Set global mutable state
 	.state('counter', 0)
 	.get('/', ({ store }) => store.counter++)
 	.listen(8080)